backend/requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.6.0
black==23.1.0
bleach==2.1.2
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.6
comm==0.1.3
cryptography==39.0.2
cycler==0.11.0
Cython==0.28.3
debugpy==1.6.7
decorator==5.1.1
docutils==0.16
entrypoints==0.2.3
executing==1.2.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.27.1
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.30
greenlet==2.0.2
gunicorn==20.1.0
html5lib==1.0.1
idna==2.10
importlib-metadata==6.1.0
importlib-resources==5.12.0
introcs==1.2.2
ipykernel==5.1.3
ipython
ipython-genutils==0.2.0
ipywidgets==7.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console
jupyter-core==4.6.0
Kivy==1.11.1
Kivy-Garden==0.1.4
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.1.2
matplotlib-inline==0.1.6
mistune==0.8.3
mypy-extensions==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==5.3.1
numpy==1.18.1
packaging==23.0
pandas==1.3.5
pandocfilters==1.4.2
parso==0.8.3
pathspec==0.11.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2023.3
pyzmq==25.0.2
qtconsole==4.3.1
regex==2022.10.31
requests==2.24.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
SQLAlchemy==1.4.46
stack-data==0.6.2
terminado==0.8.1
testpath==0.3.1
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typed-ast==1.5.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.25.10
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==3.1.0
zipp==3.15.0
unidecode==1.3.6
